refactor(json): move test models into data package
