+++ java/fory-json/src/test/java/org/apache/fory/json/data/FastContainers.java
@@ -1,0 +1,33 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.apache.fory.json.data;
+
+import java.util.Arrays;
+import java.util.List;
+import java.util.Map;
+
+public final class FastContainers {
+  public List<Boolean> booleans = Arrays.asList(Boolean.TRUE, Boolean.FALSE);
+  public Map<String, Boolean> flags = JsonTestData.flags();
+  public Map<Integer, String> intNames = JsonTestData.intNames();
+  public List<Integer> ints = Arrays.asList(1, 2);
+  public List<String> names = Arrays.asList("alpha", JsonTestData.ZH_TEXT);
+  public Map<String, Integer> scores = JsonTestData.scores();
+}
