+++ java/fory-json/src/test/java/org/apache/fory/json/data/GeneratedCollectionFields.java
@@ -1,0 +1,33 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.apache.fory.json.data;
+
+import java.util.ArrayDeque;
+import java.util.Arrays;
+import java.util.EnumSet;
+import java.util.LinkedHashSet;
+import java.util.Queue;
+import java.util.Set;
+
+public final class GeneratedCollectionFields {
+  public EnumSet<Kind> kinds = EnumSet.of(Kind.FAST, Kind.SMALL);
+  public Set<String> names = new LinkedHashSet<>(Arrays.asList("alpha", JsonTestData.ZH_TEXT));
+  public Queue<Integer> numbers = new ArrayDeque<>(Arrays.asList(1, 2));
+}
