+++ java/fory-json/src/test/java/org/apache/fory/json/data/PrivateFields.java
@@ -1,0 +1,48 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.apache.fory.json.data;
+
+public final class PrivateFields {
+  private static String staticValue = "static";
+  private transient String transientValue = "transient";
+  private int id = 11;
+  private String name = "private";
+  private String nullable;
+
+  public static int id(PrivateFields value) {
+    return value.id;
+  }
+
+  public static String name(PrivateFields value) {
+    return value.name;
+  }
+
+  public static String nullable(PrivateFields value) {
+    return value.nullable;
+  }
+
+  public static String transientValue(PrivateFields value) {
+    return value.transientValue;
+  }
+
+  public static String staticValue() {
+    return staticValue;
+  }
+}
